fix: tests with many arguments

+++ vyper/complex/create_minimal_proxy_to/create_many/main.vy
@@ -3,14 +3,14 @@ import callable as Callable
 @external
 def main(n: uint8, implementation: address) -> uint256:
     addresses: address[254] = empty(address[254])
-    for i: uint256 in range(254):
+    for i: uint8 in range(254):
         if i >= n:
             break
         addresses[i] = create_forwarder_to(implementation)
         Callable(addresses[i]).set([convert(i, uint256) + 1, convert(i, uint256) + 3], i + 2)
 
     result: uint256 = 0
-    for i: uint256 in range(254):
+    for i: uint8 in range(254):
         if i >= n:
             break
         a: uint256[2] = empty(uint256[2])
