+++ vyper/simple/algorithm/factorial.vy
@@ -43,7 +43,7 @@
 @pure
 def main(n: uint8) -> uint256:
     fact: uint256 = 1
-    for i: uint256 in range(1, 100):
+    for i: uint8 in range(1, 100):
         if not i <= n:
             break
         fact *= convert(i, uint256)
