+++ vyper/simple/loop/for/unreachable_break.vy
@@ -18,7 +18,7 @@ def entry() -> uint256:
     end: bool = True
 
     sum: uint256 = 0
-    for i: uint256 in range(1, 3):
+    for i: uint8 in range(1, 3):
         sum += i
 
         if not end:
