+++ vyper/simple/loop/range_bound/u64.vy
@@ -16,6 +16,6 @@
 @pure
 def main() -> uint64:
     sum: uint64 = 0
-    for i: uint256 in range(18446744073709551605, 18446744073709551615):
+    for i: uint64 in range(18446744073709551605, 18446744073709551615):
         sum += 1
     return sum
